microservices_architecture/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
dotenv
requests==2.28.2
beautifulsoup4==4.12.0
pydantic==1.10.7
sqlalchemy==2.0.9
psycopg2-binary==2.9.6
python-dotenv==1.0.0
apscheduler==3.10.1
loguru==0.7.0
geopy==2.3.0